modules/opentofu/pyproject.toml
[project]
name = "opentofu"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = ["dagger-io"]

[build-system]
requires = ["hatchling==1.25.0"]
build-backend = "hatchling.build"

[tool.uv.sources]
dagger-io = { path = "sdk", editable = true }
